==> src/basic_nlp_alice_zhao/1-Data Cleaning.ipynb <==
## Data Cleaning

Data Science Workflow:
1. Start with a question 
    - Goal: What makes Ali Wongs' comedy routine stand out?
2. Get & Clean the data 
    - **Input**: "How is Ali Wong different?"
            - Get: 
            - Domain expertise needed
            - Where are we going to get this data? scraps from the loft
            - How much data should we get? 
            - Which comedians? imdb (adjust the filter to get 10 comedians)
            - What time range? last 5 years
        - Data Gathering:
            - Web Scraping
                - Requests
                    - Formal: Make HTTP requests
                    - Simple: Get info from a website
                - Beautiful Soup
                    - Formal: Parse HTML documents
                    - Simple: Extract parts of a website
            - Saving Python Data:
                - Pickle
                    - Formal: Serialize Python objects
                    - Simple: Save data for later       
                    <br/>

        - Data Cleaning-format:
            1. Format#1 - Corpus
                 - a collection of texts --> We can get this from pandas dataframe
            2. Format#2 - Document-Term matrix
                 - Clean Text --> remove excess, unnecessary parts of the text
                 - Tokenize Text --> split the text into smaller pieces
                 - Document-Text Matrix - put into a matrix so a machine can read it
            3. Common data cleaning steps:
                - Make tex
                - Remove punctuation --> re: Python libraries for regular expression
                - Lowercase letters
                - Remove numbers
                - Tokenization --> You can filter stop words
     - **Ouput**: Clean, organized data in standard format 
  
  
3. Perform EDA
4. Apply Techniques
5. Share Insights
               


# import sys
# !conda install --yes --prefix {sys.prefix} plotly
# !conda install --yes --prefix {sys.prefix} -c conda-forge wordcloud
# !conda install --yes --prefix {sys.prefix} -c conda-forge textblob
# !conda install --yes --prefix {sys.prefix} -c conda-forge gensim

## Getting Data

### Data Gathering from scarps from the loft

## Web scraping
import requests
from bs4 import BeautifulSoup

## pickle imports
import dill

# Scrapes transcript data from scrapsfromtheloft.com
def url_to_transcript(url):
    ''' Returns transcript data specifically from scrapsfromtheloft.com.'''
    page = requests.get(url).text
    soup = BeautifulSoup(page, "lxml")
    text = [p.text for p in soup.find(class_="post-content").find_all('p')]
    print(url)
    return text

# URLs of transcripts in scope
urls = ['http://scrapsfromtheloft.com/2017/05/06/louis-ck-oh-my-god-full-transcript/',
        'http://scrapsfromtheloft.com/2017/04/11/dave-chappelle-age-spin-2017-full-transcript/',
        'http://scrapsfromtheloft.com/2018/03/15/ricky-gervais-humanity-transcript/',
        'http://scrapsfromtheloft.com/2017/08/07/bo-burnham-2013-full-transcript/',
        'http://scrapsfromtheloft.com/2017/05/24/bill-burr-im-sorry-feel-way-2014-full-transcript/',
        'http://scrapsfromtheloft.com/2017/04/21/jim-jefferies-bare-2014-full-transcript/',
        'http://scrapsfromtheloft.com/2017/08/02/john-mulaney-comeback-kid-2015-full-transcript/',
        'http://scrapsfromtheloft.com/2017/10/21/hasan-minhaj-homecoming-king-2017-full-transcript/',
        'http://scrapsfromtheloft.com/2017/09/19/ali-wong-baby-cobra-2016-full-transcript/',
        'http://scrapsfromtheloft.com/2017/08/03/anthony-jeselnik-thoughts-prayers-2015-full-transcript/',
        'http://scrapsfromtheloft.com/2018/03/03/mike-birbiglia-my-girlfriends-boyfriend-2013-full-transcript/',
        'http://scrapsfromtheloft.com/2017/08/19/joe-rogan-triggered-2016-full-transcript/']

# Comedian names
comedians = ['louis', 'dave', 'ricky', 'bo', 'bill', 'jim', 'john', 'hasan', 'ali', 'anthony', 'mike', 'joe']

# Actually request transcripts (takes a few minutes to run)
# transcripts = [url_to_transcript(u) for u in urls]

# # Pickle files for later use
# # Make a new directory to hold the text files
# !mkdir transcripts

# for i, c in enumerate(comedians):
#     with open("transcripts/" + c + ".txt", "wb") as file:
#         dill.dump(transcripts[i], file)

# Load pickled files
data = {}
for i, c in enumerate(comedians):
    with open("transcripts/" + c + ".txt", "rb") as file:
        data[c] = dill.load(file)

# print(data)

# Double check to make sure data has been loaded properly
data.keys()

# More checks
data['louis'][:2]

## Cleaning Data

**Common data cleaning steps on all text:**
* Make text all lower case
* Remove punctuation
* Remove numerical values
* Remove common non-sensical text (/n)
* Tokenize text
* Remove stop words

**More data cleaning steps after tokenization:**
* Stemming / lemmatization
* Parts of speech tagging
* Create bi-grams or tri-grams
* Deal with typos
* And more...

# Let's take a look at our data again
next(iter(data.keys()))

# Notice that our dictionary is currently in key: comedian, value: list of text format
next(iter(data.values()))[:1]

# We are going to change this to key: comedian, value: string format
def combine_text(list_of_text):
    '''Takes a list of text and combines them into one large chunk of text.'''
    combined_text = ' '.join(list_of_text)
    return combined_text
    
# Combine it!
data_combined = {key: [combine_text(value)] for (key, value) in data.items()}
# data_combined

# Truncate function to limit the words
def truncate_text(text, word_limit=5):
    words = text.split()[:word_limit]  # Limit to first 'word_limit' words
    return ' '.join(words) + '...'  # Add ellipsis to indicate truncation

# Combine and truncate the text in the data_combined dictionary
data_combined_5_words = {key: [truncate_text(combine_text(value))] for (key, value) in data.items()}

# Display the result
data_combined_5_words

# We can either keep it in dictionary format or put it into a pandas dataframe
import pandas as pd
pd.set_option('max_colwidth',150)

data_df = pd.DataFrame.from_dict(data_combined).transpose() ## Cool
data_df.columns = ['transcript']
data_df = data_df.sort_index()
data_df
## just "corpus"

# Let's take a look at the transcript for Ali Wong
data_df.transcript.loc['ali'][:20]

# Apply a first round of text cleaning techniques
import re
import string

def clean_text_round1(text):
    '''
    Make text lowercase, remove text in square brackets, 
    remove punctuation, and remove words containing numbers.
    '''
    text = text.lower()  # Convert to lowercase
    text = re.sub(r'\[.*?\]', '', text)  # Remove text inside square brackets
    text = re.sub(r'[%s]' % re.escape(string.punctuation), '', text)  # Remove punctuation
    text = re.sub(r'\w*\d\w*', '', text)  # Remove words containing numbers
    return text

# Let's take a look at the updated text
data_clean = pd.DataFrame(data_df["transcript"].apply(lambda text: clean_text_round1(text)))
data_clean

# Apply a second round of cleaning
def clean_text_round2(text):
    '''
    Get rid of some additional punctuation and non-sensical text that was missed the first time around.
    '''
    text = re.sub(r'[‘’“”…]', '', text)  # Remove special characters
    text = re.sub(r'\n+', ' ', text)  # Replace newline characters with spaces
    text = re.sub(r'\s+', ' ', text)  # Remove multiple spaces
    return text.strip()  # Remove leading/trailing spaces

# Let's take a look at the updated text
data_clean = pd.DataFrame(data_clean["transcript"].apply(lambda text: clean_text_round2(text)))
data_clean

# Apply a third round of cleaning

# import a lemmatization algorithm based on the built-in WordNet corpus
from nltk.stem import WordNetLemmatizer
lemmatizer = WordNetLemmatizer()

def clean_text_round3(text):
    '''
    Reducing words to its base form based on dictionry defintition
    '''
    text = ' '.join(lemmatizer.lemmatize(word) for word in text.split())
    return text.strip()  # Remove leading/trailing spaces

# Let's take a look at the updated text
data_clean = pd.DataFrame(data_clean["transcript"].apply(lambda text: clean_text_round3(text)))
data_clean


## Organize The Data
standard text formats:
1. **Corpus** - a collection of text
2. **Document-Term Matrix** - word counts in matrix format

### 1. Corpus 

# Let's take a look at our dataframe
data_df

# Let's add the comedians' full names as well
full_names = ['Ali Wong', 'Anthony Jeselnik', 'Bill Burr', 'Bo Burnham', 'Dave Chappelle', 'Hasan Minhaj',
              'Jim Jefferies', 'Joe Rogan', 'John Mulaney', 'Louis C.K.', 'Mike Birbiglia', 'Ricky Gervais']

data_df['full_name'] = full_names
data_df

# Pickle the DataFrame using dill
with open("corpus.pkl", "wb") as f:
    dill.dump(data_df, f)

### 2. Document-Term Matrix

# We are going to create a document-term matrix using CountVectorizer, and exclude common English stop words
from sklearn.feature_extraction.text import CountVectorizer

cv = CountVectorizer(stop_words='english')
data_cv = cv.fit_transform(data_clean["transcript"]) ## fit_transform() changes pd dataframes to nump() array
data_dtm = pd.DataFrame(data_cv.toarray(), columns=cv.get_feature_names_out())
data_dtm.index = data_clean.index
data_dtm

# Use dill to pickle the object
with open("dtm.pkl", "wb") as file:
    dill.dump(data_dtm, file)

# Let's also pickle the cleaned data (before we put it in document-term matrix format) and the CountVectorizer object
with open("data_clean.pkl", "wb") as f:
    dill.dump(data_clean, f)

with open("cv.pkl", "wb") as f:
    dill.dump(cv, f)
